Fix intersection bug and add Python 3.13 support (v0.7.0)
BREAKING CHANGES:
- Changed internal Real type from float to double to fix precision bug
- Binary serialization format changed (incompatible with v0.6.x)
- Updated pybind11 from v2.12.0 to v2.13.6

Bug Fix:
- Fixed critical bug where boxes with small gaps (< 1e-5) were incorrectly
  reported as intersecting due to float32 precision loss
- Root cause: PRTree constructor and insert method were using py::array_t<float>
  instead of py::array_t<Real>, causing float64 -> float32 conversion
- Example: boxes separated by 5.39e-06 would collapse to same float32 value

Changes:
- cpp/prtree.h: Changed 'using Real = float' to 'using Real = double'
- cpp/prtree.h: Updated PRTree constructor to use py::array_t<Real>
- cpp/prtree.h: Updated insert method to use py::array_t<Real>
- cpp/main.cc: Updated all pybind11 bindings to use py::array_t<double>
- third/pybind11: Updated submodule from v2.12.0 to v2.13.6
- setup.py: Bumped version to v0.7.0, added Python 3.13 support
- requirements.txt: Removed numpy<2.0 constraint for numpy 2.0 support
- .github/workflows/cibuildwheel.yml: Added Python 3.13 builds
- README.md: Added breaking changes notice and bug fix details
- tests/test_PRTree.py: Added comprehensive tests for:
  - Disjoint boxes with small gaps (regression test)
  - Touching boxes (closed interval semantics)
  - Large magnitude coordinates
  - Degenerate boxes
  - Query vs batch_query consistency

All 121 tests pass.

Fixes issue reported by Matteo Lacki.

+++ cpp/main.cc
@@ -18,7 +18,7 @@ PYBIND11_MODULE(PRTree, m)
     )pbdoc";
 
   py::class_<PRTree<T, B, 2>>(m, "_PRTree2D")
-      .def(py::init<py::array_t<T>, py::array_t<float>>(), R"pbdoc(
+      .def(py::init<py::array_t<T>, py::array_t<double>>(), R"pbdoc(
           Construct PRTree with init.
         )pbdoc")
       .def(py::init<>(), R"pbdoc(
@@ -62,7 +62,7 @@ PYBIND11_MODULE(PRTree, m)
         )pbdoc");
 
   py::class_<PRTree<T, B, 3>>(m, "_PRTree3D")
-      .def(py::init<py::array_t<T>, py::array_t<float>>(), R"pbdoc(
+      .def(py::init<py::array_t<T>, py::array_t<double>>(), R"pbdoc(
           Construct PRTree with init.
         )pbdoc")
       .def(py::init<>(), R"pbdoc(
@@ -106,7 +106,7 @@ PYBIND11_MODULE(PRTree, m)
         )pbdoc");
 
   py::class_<PRTree<T, B, 4>>(m, "_PRTree4D")
-      .def(py::init<py::array_t<T>, py::array_t<float>>(), R"pbdoc(
+      .def(py::init<py::array_t<T>, py::array_t<double>>(), R"pbdoc(
           Construct PRTree with init.
         )pbdoc")
       .def(py::init<>(), R"pbdoc(
